init renderer back/front ends

+++ engine/game.rs
@@ -7,10 +7,10 @@ pub trait Game {
     fn initialize(&mut self) -> Result<(), EngineError>;
 
     /// Update
-    fn update(&mut self, delta_time: f32) -> Result<(), EngineError>;
+    fn update(&mut self, delta_time: f64) -> Result<(), EngineError>;
 
     /// Render
-    fn render(&self, delta_time: f32) -> Result<(), EngineError>;
+    fn render(&self, delta_time: f64) -> Result<(), EngineError>;
 
     /// Resize
     fn resize(&mut self, new_width: u16, new_height: u16) -> Result<(), EngineError>;
